update sidebar file
