+++ src/app/views/sidebar/Sidebar.tsx
@@ -1,144 +1,102 @@
 import {
-  DirectionalHint,
-  FontSizes,
-  getTheme,
-  IconButton,
-  Pivot,
-  PivotItem,
-  Stack,
-  TooltipDelay,
-  TooltipHost
-} from '@fluentui/react';
-import { telemetry } from '../../../telemetry';
+  Button, ButtonProps, makeStyles, SelectTabData, SelectTabEvent, Tab, TabList, tokens
+} from '@fluentui/react-components';
+import {
+  GroupList20Regular, History20Regular, PanelLeftContract20Regular, PanelLeftExpand20Regular, Rocket20Regular
+} from '@fluentui/react-icons';
+
+import { useState } from 'react';
 import { translateMessage } from '../../utils/translate-messages';
-import SampleQueries from './sample-queries/SampleQueries';
+import History from './history/History';
+import ResourceExplorer from './resource-explorer';
 import { SampleQueriesV9 } from './sample-queries/SampleQueriesV9';
-import { sidebarStyles } from './Sidebar.styles';
-import { ResourceExplorer, History } from '../common/lazy-loader/component-registry';
-interface ISidebar {
-  currentTab: string;
-  setSidebarTabSelection: Function;
-  showSidebar: boolean;
-  toggleSidebar: Function;
-  mobileScreen: boolean;
+
+interface IShowSidebar {
+  show: boolean
+  handleShow: ()=>void
+}
+const useStyles = makeStyles({
+  container: {
+    display: 'flex',
+    flexDirection: 'column',
+    overflow: 'hidden',
+    height: '100%',
+    padding: `0 ${tokens.spacingHorizontalS}`,
+    backgroundColor: tokens.colorNeutralBackground2,
+    borderRightWidth: tokens.strokeWidthThin,
+    borderRight: `1px solid ${tokens.colorNeutralForeground3}`
+  },
+  sidebarToggle: {
+    marginLeft: 'auto'
+  }
+})
+const SidebarToggle = (props: IShowSidebar & ButtonProps)=>{
+  const {show, handleShow} = props;
+  const PanelIcon = ()=> show ? <PanelLeftContract20Regular/>: <PanelLeftExpand20Regular/>
+
+  return <Button appearance='subtle' icon={PanelIcon()} onClick={handleShow} {...props}></Button>
 }
 
-export const Sidebar = (props: ISidebar) => {
-  const theme = getTheme();
-  const styles = sidebarStyles(theme).sidebarButtons;
+interface SidebarProps {
+  handleToggleSelect: (showSidebarValue: boolean) => void;
+}
+const Sidebar = (props: SidebarProps)=>{
+  const sidebarStyles = useStyles();
+  const [showSidebarValue, setShowSidebarValue] = useState(true);
+  const [selectedValue, setSelectedValue] = useState<string>('sample-queries');
 
-  const onPivotItemClick = (item?: PivotItem) => {
-    if (!item) { return; }
-    const key = item.props.itemKey;
-    if (key) {
-      props.setSidebarTabSelection(key);
-      telemetry.trackTabClickEvent(key);
+  const onTabSelect = (_event: SelectTabEvent, data: SelectTabData) => {
+    setSelectedValue(data.value as string);
+    if (!showSidebarValue) {
+      handleShow();
     }
-  }
-  const openComponent = (key: string) => {
-    props.toggleSidebar();
-    props.setSidebarTabSelection(key);
+  };
+
+  const handleShow = () => {
+    const newState = !showSidebarValue;
+    setShowSidebarValue(newState);
+    props.handleToggleSelect(newState);
+  };
+
+  const tabItems: Record<string, JSX.Element> = {
+    'sample-queries': <SampleQueriesV9 />,
+    'resources': <ResourceExplorer />,
+    'history': <History />
   }
 
   return (
-    <div>
-      {props.showSidebar &&
-        <Pivot onLinkClick={onPivotItemClick}
-          overflowBehavior='menu'
-          overflowAriaLabel={translateMessage('More items')}
-          defaultSelectedKey={props.currentTab}
-          styles={{ text: { fontSize: FontSizes.size14 } }}>
-          <PivotItem
-            headerText={translateMessage('Sample Queries')}
-            itemIcon='Rocket'
-            itemKey='sample-queries'
-            headerButtonProps={{
-              'aria-controls': 'sample-queries-tab'
-            }}
-          >
-            {/* <div id={'sample-queries-tab'}><SampleQueries /></div> */}
-            <div id={'sample-queries-tab'}><SampleQueriesV9 /></div>
-
-          </PivotItem>
-          <PivotItem
-            headerText={translateMessage('Resources')}
-            itemIcon='ExploreData'
-            itemKey='resources'
-            headerButtonProps={{
-              'aria-controls': 'resources-tab'
-            }}
-          >
-            <div id={'resources-tab'}><ResourceExplorer /></div>
-          </PivotItem>
-          <PivotItem
-            headerText={translateMessage('History')}
-            itemIcon='History'
-            itemKey='history'
-            headerButtonProps={{
-              'aria-controls': 'history-tab'
-            }}
-          >
-            <div id={'history-tab'}><History /></div>
-          </PivotItem>
-        </Pivot>
-      }
-      {!props.showSidebar && !props.mobileScreen && (
-        <Stack tokens={{ childrenGap: 10 }}>
-          <TooltipHost
-            content={
-              <div style={{ padding: '3px' }}>
-                {translateMessage('Sample Queries')}
-              </div>}
-            calloutProps={{ gapSpace: 0 }}
-            directionalHint={DirectionalHint.bottomCenter}
-            styles={{ root: { display: 'inline-block' } }}
-            delay={TooltipDelay.zero}
-          >
-            <IconButton
-              iconProps={{ iconName: 'Rocket' }}
-              ariaLabel={translateMessage('Sample Queries')}
-              onClick={() => openComponent('sample-queries')}
-              styles={styles}
-            />
-          </TooltipHost>
-          <TooltipHost
-            content={
-              <div style={{ padding: '3px' }}>
-                {translateMessage('Resources')}
-              </div>}
-            calloutProps={{ gapSpace: 0 }}
-            directionalHint={DirectionalHint.bottomCenter}
-            styles={{ root: { display: 'inline-block' } }}
-            delay={TooltipDelay.zero}
-          >
-            <IconButton
-              iconProps={{ iconName: 'ExploreData' }}
-              ariaLabel={translateMessage('Resources')}
-              onClick={() => openComponent('resources')}
-              styles={styles}
-            />
-          </TooltipHost>
-          <TooltipHost
-            content={
-              <div style={{ padding: '3px' }}>
-                {translateMessage('History')}
-              </div>}
-            calloutProps={{ gapSpace: 0 }}
-            directionalHint={DirectionalHint.bottomCenter}
-            styles={{ root: { display: 'inline-block' } }}
-            delay={TooltipDelay.zero}
-          >
-            <IconButton
-              iconProps={{ iconName: 'History' }}
-              ariaLabel={translateMessage('History')}
-              onClick={() => openComponent('history')}
-              styles={styles}
-            />
-          </TooltipHost>
-        </Stack>)
-      }
+    <div className={sidebarStyles.container}>
+      <SidebarToggle className={sidebarStyles.sidebarToggle} show={showSidebarValue} handleShow={handleShow}/>
+      <TabList
+        selectedValue={selectedValue} onTabSelect={onTabSelect} size='large' vertical>
+        {renderTablistItems(showSidebarValue)}
+      </TabList>
+      <div>
+        {showSidebarValue && tabItems[selectedValue]}
+      </div>
     </div>
-  );
-};
+  )
+}
 
+const renderTablistItems = (showSidebar: boolean) =>{
+  if (showSidebar) {
+    return (
+      <>
+        <Tab id='sample-queries' value='sample-queries' icon={<Rocket20Regular />}>
+          {translateMessage('Sample Queries')}</Tab>
+        <Tab id='resources' value='resources' icon={<GroupList20Regular />}>{translateMessage('Resources')}</Tab>
+        <Tab id='history' value='history' icon={<History20Regular />}>{translateMessage('History')}</Tab>
+      </>
+    )
+  }
+  return (
+    <>
+      <Tab id='sample-queries' value='sample-queries' icon={<Rocket20Regular />}></Tab>
+      <Tab id='resources' value='resources' icon={<GroupList20Regular />}></Tab>
+      <Tab id='history' value='history' icon={<History20Regular />}></Tab>
+    </>
+  )
+}
+
+export { Sidebar };
